Parses peerlist, tries to connect to peer.
Finished making an array of Peer objects from the tracker's compact
response.

The "other_client" var on line 160 should define the other client
you're running so you can try handshakes. I attempted with utorrent, to
no avail...

The included torrent file is another frostwire torrent with fewer than
10 peers, which made it easier to go through the peerlists. Given the
current state of the code though, it might not be necessary. I've tried
to cut down on the output the program generates for debug values etc.
that I'm fairly confident about, but I've left in puts statements for
debugging here and there.